rename render to view, visualize_stl to view_stl
