+++ docs/api.md
@@ -2,14 +2,14 @@
 title: API Reference
 ---
 
-## ::: jupyterscad.render
+## ::: jupyterscad.view
+    rendering:
+      show_root_full_path: false
+
+## ::: jupyterscad.view_stl
     rendering:
       show_root_full_path: false
 
 ## ::: jupyterscad.render_stl
-    rendering:
-      show_root_full_path: false
-
-## ::: jupyterscad.visualize_stl
     rendering:
       show_root_full_path: false
