+++ docs/usage.md
@@ -19,36 +19,30 @@ Note: If a 3D object description integrates an external stl file, then the stl m
 
 See the [OpenSCAD language](https://en.wikibooks.org/wiki/OpenSCAD_User_Manual#The_OpenSCAD_Language_Reference) and [SolidPython2](https://github.com/jeff-dh/SolidPython) pages for more information on how to use these tools.
 
-## Rendering an OpenSCAD object
+## Rendering and Viewing an OpenSCAD object
 
 ### Visualizing and Rendering to an stl
 
-To visualize an OpenSCAD object, use `render`:
+To visualize an OpenSCAD object, use `view`:
+
 ```python
-from jupyterscad import render
+from jupyterscad import view
 from solid2 import cube
 
 obj = cube([1.5,2,1],center=True)
-render(obj)
+view(obj)
 ```
 ![render example](https://github.com/jreiberkyle/jupyterscad/blob/main/images/render_cube.png?raw=True)
 
-The rendering can also save the stl file by defining 'outfile':
+The rendered 3D OpenSCAD object can also save the stl file by defining 'outfile':
 
 ```python
-from jupyterscad import render
+from jupyterscad import view
 from solid2 import cube
 
 obj = cube([1.5,2,1],center=True)
-render(obj, outfile='obj.stl')
+view(obj, outfile='obj.stl')
 ```
-
-### Adjusting grid size
-
-In `render` and `visualize_stl`, the default grid unit is 1. The grid unit can
-be set to e.g. `10` with `grid_unit=10`, disabled with `grid_unit=0` or set to
-automatic scaling with `grid_unit=-1`.
-
 
 ### Rendering directly to an stl
 
@@ -64,7 +58,13 @@ render_stl(obj, 'obj.stl')
 A stl can be visualized with:
 
 ```python
-from jupyterscad import visualize_stl
+from jupyterscad import view_stl
 
-visualize_stl('obj.stl')
+view_stl('obj.stl')
 ```
+
+### Adjusting grid size
+
+In `view` and `view_stl`, the default grid unit is 1. The grid unit can
+be set to e.g. `10` with `grid_unit=10`, disabled with `grid_unit=0` or set to
+automatic scaling with `grid_unit=-1`.
