+++ src/jupyterscad/__init__.py
@@ -14,7 +14,7 @@ PARTICULAR PURPOSE. See the GNU General Public License for more details.
 You should have received a copy of the GNU General Public License along with
 this program. If not, see <https://www.gnu.org/licenses/>.
 """
-from ._render import render, render_stl
-from ._visualize import visualize_stl
+from ._render import render_stl
+from ._view import view, view_stl
 
-__all__ = ["render", "render_stl", "visualize_stl"]
+__all__ = ["render_stl", "view", "view_stl"]
